requirements.txt
google-cloud-bigquery
google-api-core
pandas
pyarrow
python-dotenv
PyYAML
streamlit